Add ClientServerExample source files
When testing, please pay attention to the following:
1. Make sure that you start the Server before connecting any of the
clients. The client doesn´'t check beforehand if there is a socket
listening on the other end, so the code will throw an error.
2. The "Stop Server" button on the Server side will not work(TO DO)
3. Messages from the server side are sent to all connected clients. When
more messages are received (on either the client or the server app) you
may not notice the new messages because of the scroll bar. Just scroll
down to see the latest messages.

!! Some functionalities may be misssing and you'll find many bugs while
testing. This is a Client-Server communication example, the goal was in
the first place to ensure a good communication between the Client and
the Server app. We'll figure out together the rest.